Create README.md
In this project, I introduce the cost function to logistic regression and Random Forest classification in order to create a cost-sensitive model. I figure out two ways to introduce the cost function to the original machine learning models and compare the two ways in terms of model performance and total costs.